requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
openai
pytest
httpx<0.28